DISPATCH-390: Get rid of ssl_profile reference counts
Copy SSL profile date at the time the connector or listener is configured, so
that it is safe to delete an SSL profile even if there are connectors/listeners
that used it.

Simplifying the lifecycle, getting rid of reference counts and removing the user
restrictions on removing SSL profiles seems to justify the small memory
duplication for the SSL profile data.

This means you cannot change an SSL profile and have the changes applied on the
next retry of a connector, but that seems like it would be an insane and
thread-unsafe thing to do anyway.